refactor: change logic on cli command and migrate to cobra cli
